requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
g4f==0.3.8.3
